Add "persist_media_on_sort_change" option to config file
As stated in the config file, this allows the user to disable the super-fancy flash reduction technology that keeps the same image visible when the sort method changes. I'll elaborate a bit about this.

In TagViewer 1, I was faced with an issue: when the sort method changed, the image would change. This is because the media number stayed the same, so the actual item that it referenced would change, causing the visible item to change. I decided that this was undesirable and wrote the bodgiest code of my life to fix it.

This type of unexpected and unpreventable updates was the reason I wrote [StateMan](https://github.com/mattfbacon/stateman), TagViewer 2's state manager, to be explicit in its dependencies. It's already made my development process easier with the dark mode code that needed to know whether the app was in fullscreen and vice versa (will be committed soon).

That StateMan is so incredibly simple and intuitive to use (in my opinion) is the only reason I'm able to include this option. But enough sidebar.

If this option is set to True, TagViewer 2 will behave like TagViewer 1, changing the media number when the sort method changes. If it's false, it will not update the media number, causing the visible media to change but the media number to stay the same.

I briefly considered adding a third option to keep both the visible media and the media number the same by not updating the visible media at all when the sort method changes. The image would only update when the user continued navigating between images. However, I quickly dismissed it as confusing, even misleading, to the user. If it becomes clear that this third option is desirable, I will not hesitate to add it. In fact, it will be the easiest of the three options to implement!