Don't use ChainMap for config
The ChainMap approach was good because it prevented unnecessary items in the user's config/cache files. However, Python's treatment of nested dictionary updates causes the defaults dictionary to be updated instead of user's config when a dictionary doesn't exist. This is because Python first checks if the dictionary exists in the user's config. Since it doesn't it moves to the defaults, finds it, and sets one of its properties.

It would be incredibly janky to solve this with empty dictionaries and stuff, and the only other alternative is flattening the config/cache, so I made the decision to remove this slight convenience and just copy the full default config file into the user's config.