Add TagSpace defaults settings pages
First of all, I split tags and props into two pages, because it was confusing to have two TreeViews on one pages.
Second, for now I implemented the color choice as a CellRendererText, and left a TODO comment, because I still haven't figured out how to put a widget in a CellRenderer so I'm going to leave it for later and put in a worse but working solution so I can move on.